Add more verbose logging in case of error and fix bug
Add a more verbose message when an override file cannot be found for easier debugging.

Also, it is a valid scenario where there are charts but no override files for them, so
if we do not find an override file for a chart, we should not `exit 1`. Instead just
pring a message and move on.